+++ .eslintrc.js
@@ -16,6 +16,16 @@ module.exports = {
   'plugins': [
     'jsdoc'
   ],
+  'rules': {
+    'max-lines-per-function': [
+      'warn',
+      {
+        'max': 50,
+        'skipBlankLines': true,
+        'skipComments': true
+      }
+    ]
+  },
   'settings': {
     'jsdoc': {
       'tagNamePreference': {
